Tarea No. 8 Cotizaciones

+++ RegistroTecnicos/Services/TrabajoServices.cs
@@ -1,27 +1,23 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using RegistroTecnicos.DAL;
 using RegistroTecnicos.Models;
+using System.Drawing;
 using System.Linq.Expressions;
 
 namespace RegistroTecnicos.Services
 {
-    public class TrabajoServices
+    public class TrabajoServices(IDbContextFactory<Context> DbFactory)
     {
         private readonly Context _context;
-
-        public TrabajoServices(Context context)
-        {
-            _context = context;
-        }
-
         public async Task<bool> Existe(int trabajoId)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             return await _context.Trabajos.AnyAsync(t => t.TrabajoId == trabajoId);
         }
 
         public async Task<bool> Insertar(Trabajos trabajo)
         {
-            
+            await using var _context = await DbFactory.CreateDbContextAsync();
             foreach (var detalle in trabajo.trabajosDetalle)
             {
                 var articulo = await BuscarArticulos(detalle.ArticuloId);
@@ -37,24 +33,23 @@ namespace RegistroTecnicos.Services
                 }
                 else
                 {
-                   
                     return false;
                 }
             }
             _context.Trabajos.Add(trabajo);
-            await _context.SaveChangesAsync();
-            return true;
+            return await _context.SaveChangesAsync() > 0;
         }
 
         private async Task<bool> Modificar(Trabajos trabajo)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             _context.Update(trabajo);
-            return await _context
-                .SaveChangesAsync() > 0;
+            return await _context.SaveChangesAsync() > 0;
         }
 
         public async Task<Trabajos?> Buscar(int trabajoId)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             return await _context.Trabajos
                 .Include(t => t.trabajosDetalle)
                 .AsNoTracking()
@@ -63,6 +58,7 @@ namespace RegistroTecnicos.Services
 
         public async Task<List<Trabajos>> Listar(Expression<Func<Trabajos, bool>> criterio)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             return await _context.Trabajos
                 .Include(t => t.trabajosDetalle)
                 .AsNoTracking()
@@ -72,6 +68,7 @@ namespace RegistroTecnicos.Services
 
         public async Task<List<Articulos>> GetArticulos()
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             return await _context.Articulos
                 .AsNoTracking()
                 .ToListAsync();
@@ -79,6 +76,7 @@ namespace RegistroTecnicos.Services
 
         public async Task<Articulos> BuscarArticulos(int id)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             return await _context.Articulos
                 .AsNoTracking()
                 .FirstOrDefaultAsync(a => a.ArticuloId == id);
@@ -86,6 +84,7 @@ namespace RegistroTecnicos.Services
 
         public async Task<bool> ActualizarArticulo(Articulos articulo)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             _context.Articulos.Update(articulo);
             return await _context
                 .SaveChangesAsync() > 0;
@@ -103,6 +102,7 @@ namespace RegistroTecnicos.Services
 
         public async Task<List<TrabajosDetalle>> BuscarTrabajoDetalle(int trabajoId)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             return await _context.trabajosDetalles
                 .Include(a => a.Articulos)
                 .Where(t => t.TrabajoId == trabajoId)
@@ -112,6 +112,7 @@ namespace RegistroTecnicos.Services
 
         public async Task<bool> Eliminar(int trabajoId)
         {
+            await using var _context = await DbFactory.CreateDbContextAsync();
             var detalles = await BuscarTrabajoDetalle(trabajoId);
 
             foreach (var detalle in detalles)
@@ -123,15 +124,25 @@ namespace RegistroTecnicos.Services
                     await ActualizarArticulo(articulo);
                 }
             }
+            var cobro = await _context.Trabajos
+                   .Include(c => c.trabajosDetalle)
+                   .FirstOrDefaultAsync(t => t.TrabajoId == trabajoId);
+            if (cobro == null) return false;
 
-            var trabajos = await _context.Trabajos
-                .Where(t => t.TrabajoId == trabajoId)
-                .ExecuteDeleteAsync();
-            return trabajos > 0;
+            _context.trabajosDetalles.RemoveRange(cobro.trabajosDetalle);
+            _context.Trabajos.Remove(cobro);
+            var cantidad = await _context.SaveChangesAsync();
+            return cantidad > 0;
+            /* var trabajos = await _context.Trabajos
+                 .Where(t => t.TrabajoId == trabajoId)
+                 .ExecuteDeleteAsync();
+             return trabajos > 0;*/
         }
 
         public async Task<List<TrabajosDetalle>> ListarTrabajoDetalle(int trabajoId)
         {
+
+            await using var _context = await DbFactory.CreateDbContextAsync();
             var detalle = await _context.trabajosDetalles
                 .Where(d => d.TrabajoId == trabajoId)
                 .ToListAsync();
